+++ DjangoLibrary/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from robot.api import logger
 from robot.libraries.BuiltIn import BuiltIn
+from warnings import warn
 
 import base64
 import os
@@ -34,7 +35,7 @@ class DjangoLibrary:
 
     def __init__(self, host="0.0.0.0", port=8000, path='mysite/mysite',
                  manage='mysite/manage.py', settings='mysite.settings',
-                 db="test.db"):
+                 db=None):
         """Django2Library can be imported with optional arguments.
 
         `host` is the hostname of your Django instance. Default value is
@@ -49,8 +50,6 @@ class DjangoLibrary:
 
         `settings` is the path to your Django instance settings.py.
 
-        `db` is the path to your Django instance database.
-
         Examples:
         | Library | Selenium2Library | timeout=15        | implicit_wait=0.5  | # Sets default timeout to 15 seconds and the default implicit_wait to 0.5 seconds. |  # noqa
         | Library | DjangoLibrary    | 127.0.0.1         | 55001              | path=mysite/mysite | manage=mysite/manage.py | settings=mysite.settings | db=mysite/db.sqlite3 | # Sets default hostname to 127.0.0.1 and the default port to 55001.                |  # noqa
@@ -60,38 +59,50 @@ class DjangoLibrary:
         self.path = os.path.realpath(path)
         self.manage = os.path.realpath(manage)
         self.settings = settings
-        self.db = os.path.realpath(db)
+        if db:
+            warn(
+                "Using the DjangoLibrary 'db' parameter is deprecated. " +
+                "Use the 'settings' parameter instead to set a " +
+                "database connection."
+            )
 
-    def clear_db(self):
-        """Clear the Django default database by running
-        'python manage.py migrate'.
-        """
-        # XXX: Flush seems to be not working
-        # args = [
-        #     'python',
-        #     'mysite/manage.py',
-        #     'flush',
-        #     '--noinput',
-        # ]
-        args = [
-            'rm',
-            self.db,
-        ]
-        subprocess.call(args)
-        import django
-        try:
-            django.__version__
-            migrate_stmt = 'migrate'
-        except AttributeError:
-            migrate_stmt = 'syncdb'
+    def manage_makemigrations(self):
+        """Create migration by running 'python manage.py makemigrations'."""
         args = [
             'python',
             self.manage,
-            migrate_stmt,
-            '--noinput',
-            '--settings=%s' % self.settings,
+            'makemigrations',
         ]
         subprocess.call(args)
+
+    def manage_migrate(self):
+        """Execute migration by running 'python manage.py migrate'."""
+        args = [
+            'python',
+            self.manage,
+            'migrate',
+        ]
+        subprocess.call(args)
+
+    def manage_flush(self):
+        """Clear database by running 'python manage.py flush'."""
+        args = [
+            'python',
+            self.manage,
+            'flush',
+            '--noinput',
+        ]
+        subprocess.call(args)
+
+    def clear_db(self):
+        """Clear database. This is a legacy keyword now. Use 'Manage Flush'
+           instead.
+        """
+        warn(
+            "The DjangoLibrary 'clear_db' keyword is deprecated. " +
+            "Use the 'manage_flush' keyword instead."
+        )
+        self.manage_flush()
 
     def create_user(self, username, email, password, **kwargs):
         """Create a regular Django user in the default auth model."""
@@ -136,6 +147,8 @@ user.save()""" % {
     def start_django(self):
         """Start the Django server."""
         self.clear_db()
+        self.manage_makemigrations()
+        self.manage_migrate()
         logger.console("-" * 78)
         args = [
             'python',
@@ -223,21 +236,3 @@ user.save()""" % {
         selenium2lib = BuiltIn().get_library_instance('Selenium2Library')
         selenium2lib.execute_javascript(
             "document.cookie = 'autologin=;path=/;domain=localhost;';")
-
-    def pause(self):
-        """Visually pause test execution with interactive dialog by importing
-        **Dialogs**-library and calling its **Pause Execution**-keyword.
-        """
-        from robotframework.libraries.Dialogs import pause_execution
-        pause_execution()
-
-    def debug(self):
-        """Pause test execution with interactive debugger (REPL) in the
-        current shell.
-
-        This keyword is based on ``roboframework-debuglibrary``
-        and requires that the used Python is compiled with
-        ``readline``-support.
-        """
-        from robotframework_debuglibrary.DebugLibrary import Debug
-        Debug()